+++ Forge/changelog.md
@@ -5,3 +5,5 @@
 * Added Strawberry Seeds
 * Added Sugar Beets food properties
 * Added Advancement for Corn and Corn Seeds
+* Added Jar recipe
+* Added Honey Jar recipe